Extract SidebarContent to each view
