+++ src/css/LiveSplit.scss
@@ -1,4 +1,5 @@
 @import 'mobile';
+@import 'Sidebar';
 @import 'variables';
 
 $sidebar-button-height: 42px;
